+++ apps/docs/src/examples/animation/transitions/Scene.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { MathUtils } from 'three'
   import { T } from '@threlte/core'
   import Character from './Character.svelte'
 
@@ -14,11 +15,17 @@
 />
 
 <T.AmbientLight />
-<T.DirectionalLight position={[10, 5, 5]} />
+<T.DirectionalLight
+  position={[10, 5, 5]}
+  castShadow
+/>
 
 <Character actionKey={action} />
 
-<T.Mesh rotation.x={-90 * (Math.PI / 180)}>
+<T.Mesh
+  rotation.x={MathUtils.degToRad(-90)}
+  receiveShadow
+>
   <T.CircleGeometry args={[3, 72]} />
   <T.MeshStandardMaterial color={'white'} />
 </T.Mesh>