+++ apps/docs/src/examples/extras/soft-shadows/Scene.svelte
@@ -7,7 +7,7 @@
 
   const { scene } = useThrelte()
 
-  let lightHelper: DirectionalLightHelper
+  let lightHelper = $state<DirectionalLightHelper>()
 </script>
 
 <T.PerspectiveCamera