+++ apps/docs/src/examples/geometry/random-placement/basic-random/assets/bush.svelte
@@ -27,7 +27,10 @@
 </script>
 
 {#await assets then [$gltf, $texture1]}
-  <InstancedMesh>
+  <InstancedMesh
+    castShadow
+    receiveShadow
+  >
     <T is={$gltf.nodes.Bush.geometry} />
     <T.MeshStandardMaterial
       map={$texture1}