+++ apps/docs/src/examples/geometry/random-placement/variable-poisson-random/assets/rock.svelte
@@ -24,7 +24,10 @@
 </script>
 
 {#if $gltf}
-  <InstancedMesh>
+  <InstancedMesh
+    castShadow
+    receiveShadow
+  >
     <T is={$gltf.nodes.Rock_2.geometry} />
     <T.MeshStandardMaterial color="grey" />
     {#each transformData as randomValues}