+++ apps/docs/src/examples/tutorials/animating-a-spaceship/Stars.svelte
@@ -1,54 +1,63 @@
-<script>
+<script lang="ts">
   import { T, useTask } from '@threlte/core'
   import { Instance, InstancedMesh, useTexture } from '@threlte/extras'
-  import { Color, DoubleSide, Vector3 } from 'three'
+  import { Color, DoubleSide, MathUtils, type Vector3Tuple } from 'three'
 
   let STARS_COUNT = 350
-  let colors = ['#fcaa67', '#C75D59', '#ffffc7', '#8CC5C6', '#A5898C']
-  let stars = []
+  let colors = ['#fcaa67', '#C75D59', '#ffffc7', '#8CC5C6', '#A5898C'] as const
+  let stars = $state<Star[]>([])
 
   const map = useTexture('/spaceship-tutorial/textures/star.png')
 
-  function r(min, max) {
+  function r(min: number, max: number): number {
     let diff = Math.random() * (max - min)
     return min + diff
   }
 
-  function resetStar(star) {
+  interface Star {
+    id: string
+    position: Vector3Tuple
+    length: number
+    speed: number
+    color: Color
+  }
+
+  function resetStar(star: Star) {
     if (r(0, 1) > 0.8) {
-      star.pos = new Vector3(r(-10, -30), r(-5, 5), r(6, -6))
-      star.len = r(1.5, 15)
+      star.position = [r(-10, -30), r(-5, 5), r(6, -6)]
+      star.length = r(1.5, 15)
     } else {
-      star.pos = new Vector3(r(-15, -45), r(-10.5, 1.5), r(30, -45))
-      star.len = r(2.5, 20)
+      star.position = [r(-15, -45), r(-10.5, 1.5), r(30, -45)]
+      star.length = r(2.5, 20)
     }
 
     star.speed = r(19.5, 42)
-    star.rad = r(0.04, 0.07)
-    star.color = new Color(colors[Math.floor(Math.random() * colors.length)])
+    star.color
+      .set(colors[Math.floor(Math.random() * colors.length)] ?? 'white')
       .convertSRGBToLinear()
       .multiplyScalar(1.3)
-
-    return star
   }
 
   for (let i = 0; i < STARS_COUNT; i++) {
-    let star = {
-      pos: null,
-      len: null,
-      speed: null,
-      color: null
+    const star: Star = {
+      id: MathUtils.generateUUID(),
+      position: [0, 0, 0],
+      length: 0,
+      speed: 0,
+      color: new Color()
     }
 
-    stars.push(resetStar(star))
+    resetStar(star)
+    stars.push(star)
   }
 
   useTask((delta) => {
-    stars.forEach((star) => {
-      star.pos.x += star.speed * delta
-      if (star.pos.x > 40) resetStar(star)
-    })
-    stars = stars
+    for (const star of stars) {
+      star.position[0] += star.speed * delta
+      if (star.position[0] > 40) {
+        resetStar(star)
+      }
+    }
   })
 </script>
 
@@ -64,11 +73,11 @@
       transparent
     />
 
-    {#each stars as star}
+    {#each stars as { id, position, length, color } (id)}
       <Instance
-        position={[star.pos.x, star.pos.y, star.pos.z]}
-        scale={[star.len, 1, 1]}
-        color={star.color}
+        {position}
+        scale={[length, 1, 1]}
+        {color}
       />
     {/each}
   </InstancedMesh>