+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @threlte/core
 
+## 8.1.5
+
+### Patch Changes
+
+- 64040a5: Fix typechecking for rest params on T components
+
 ## 8.1.4
 
 ### Patch Changes