+++ packages/flex/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @threlte/flex
 
+## 2.0.6
+
+### Patch Changes
+
+- 6ef7086: Improve internal types
+
 ## 2.0.5
 
 ### Patch Changes